Unify test/environment.py with test/env.py
